+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-Last change: 10-MAY-2026 CEST
+Last change: 11-MAY-2026 CEST
 
 ## 0.10.1
 
@@ -12,6 +12,7 @@ Last change: 10-MAY-2026 CEST
 - Render: fixed panning in 3D
 - Render: manual setting of scale bar switches off automatic scale bar length
 - Render: faster non-circle picking by smarter indexing
+- Filter: apply filtering steps from metadata
 
 ## 0.10.0